Move Jenkins Resource Request to separate Argo ApplicationSet
The Jenkins Promise introduced in the `kratix-platform-resources`
Argo ApplicationSet creates the CRD that this Resource Request
depends on. If they are done in the same ApplicationSet, Argo will
never be able to sync them.

This is also likely a different usecase for Kratix. As in, the idea
behind a Promise and a Resource Request is that these are likely
two different personas. The Promise is introduced by a team/engineer
who offers something as a service, while a Resource Request is made
by a consumer of that service. This might map to a platform engineer
creating the Promise and an application engineer making the Resource
Request, but it could be a platform team offering something to another
Platform team.